Debug options toggle added to Settings; ConnectionOptions added into UI and RecentStreamsDataStore
