+++ interactiveplayer/src/main/java/io/dolby/interactiveplayer/datastore/RecentStreamsDataStore.kt
@@ -1,14 +1,15 @@
 package io.dolby.interactiveplayer.datastore
 
+import io.dolby.interactiveplayer.rts.domain.ConnectOptions
 import io.dolby.interactiveplayer.rts.domain.StreamingData
 import kotlinx.coroutines.flow.Flow
 
 interface RecentStreamsDataStore {
     var recentStreams: Flow<List<StreamDetail>>
 
-    suspend fun recentStream(streamName: String): StreamDetail?
+    suspend fun recentStream(accountId: String, streamName: String): StreamDetail?
 
-    fun addStreamDetail(streamingData: StreamingData)
+    fun addStreamDetail(streamingData: StreamingData, connectOptions: ConnectOptions)
 
     suspend fun clearAll()
