+++ interactiveplayer/src/main/java/io/dolby/interactiveplayer/navigation/Screen.kt
@@ -3,19 +3,16 @@ package io.dolby.interactiveplayer.navigation
 import io.dolby.interactiveplayer.rts.domain.StreamingData
 
 sealed class Screen(val route: String) {
-    object DetailInputScreen : Screen(route = "login") {
-        const val ARG_STREAM_NAME_TO_PLAY = "streamNameToPlay"
-        const val ARG_ACCOUNT_ID_TO_PLAY = "accountIDToPlay"
-        const val ARG_USE_DEV_ENV = "useDevEnv"
-        const val ARG_DISABLE_AUDIO = "disableAudio"
-        const val ARG_RTC_LOGS = "rtcLogs"
-        const val ARG_VIDEO_JITTER = "videoJitter"
-    }
+    object DetailInputScreen : Screen(route = "login")
 
     object MultiStreamingScreen :
         Screen(route = "multistreaming/streamName={streamName}&accountId={accountId}") {
         const val ARG_STREAM_NAME = "streamName"
         const val ARG_ACCOUNT_ID = "accountId"
+        const val ARG_USE_DEV_ENV = "useDevEnv"
+        const val ARG_DISABLE_AUDIO = "disableAudio"
+        const val ARG_RTC_LOGS = "rtcLogs"
+        const val ARG_VIDEO_JITTER = "videoJitter"
         fun route(model: StreamingData): String {
             val streamName = model.streamName
             val accountId = model.accountId
