+++ interactiveplayer/src/main/java/io/dolby/interactiveplayer/preferenceStore/PrefsStore.kt
@@ -1,6 +1,5 @@
 package io.dolby.interactiveplayer.preferenceStore
 
-import androidx.datastore.preferences.core.Preferences
 import io.dolby.interactiveplayer.rts.domain.StreamingData
 import kotlinx.coroutines.flow.Flow
 
@@ -28,64 +27,3 @@ interface PrefsStore {
     suspend fun clear(streamingData: StreamingData)
     suspend fun clearAllStreamSettings()
 }
-interface Prefs {
-    val data: Flow<Preferences>
-
-    val showSourceLabels: Flow<Boolean>
-    val multiviewLayout: Flow<MultiviewLayout>
-    val streamSourceOrder: Flow<StreamSortOrder>
-    val audioSelection: Flow<AudioSelection>
-
-    suspend fun updateShowSourceLabels(show: Boolean)
-    suspend fun updateMultiviewLayout(layout: MultiviewLayout)
-    suspend fun updateStreamSourceOrder(order: StreamSortOrder)
-    suspend fun updateAudioSelection(selection: AudioSelection)
-
-    suspend fun clearPreference()
-}
-
-enum class MultiviewLayout {
-    ListView, SingleStreamView, GridView;
-
-    fun stringResource(): Int {
-        return when (this) {
-            ListView -> R.string.settings_multiview_layout_list_view
-            SingleStreamView -> R.string.settings_multiview_layout_single_view
-            GridView -> R.string.settings_multiview_layout_grid_view
-        }
-    }
-    companion object {
-        fun valueToLayout(value: String): MultiviewLayout {
-            return try {
-                MultiviewLayout.valueOf(value)
-            } catch (ex: Exception) {
-                default
-            }
-        }
-        val default: MultiviewLayout = ListView
-    }
-}
-
-enum class StreamSortOrder {
-    ConnectionOrder, AlphaNumeric;
-
-    fun stringResource(): Int {
-        return when (this) {
-            ConnectionOrder -> R.string.settings_stream_sort_order_connection
-            AlphaNumeric -> R.string.settings_stream_sort_order_alphanumeric
-        }
-    }
-
-    companion object {
-        fun valueToSortOrder(value: String): StreamSortOrder {
-            return try {
-                StreamSortOrder.valueOf(value)
-            } catch (ex: Exception) {
-                default
-            }
-        }
-
-        val default: StreamSortOrder = ConnectionOrder
-    }
-}
-
